[molecule][collectd_connection] Change retry criteria for checking network
The retry module is used for checking whether the network
plugin is set up correctly in order to compensate for
delays in generating, sending and receiving metrics.

The task parameters were updated so that the failure happend
with a non-zero rc, and the task continues to retry until
there is a list of metrics (i.e. stdout_lines|len != 0), or
the retry counter runs out.
Previously, the retry counter wouldn't get past the first
iteration due to incorrect failure criteria.